backend/requirements.txt
Flask
Flask-Cors
python-dotenv
requests
PyMuPDF
selenium
webdriver-manager
unidecode
sendgrid